check modifiers before stash calling in _hdlr_entries.
mt:entries invokes $ctx->stash for prepopulated entries. 
stash invoking takes high costs, because it will invoke the MT::Promise object's force method, if it is. but, if mt:entries was called with some of modifiers, it discards the prepopulated entries after invoked for stash.
this change will make that to check the modifiers before stash invoking, and prevents an unnecessary force of MT::Promise objects.


git-svn-id: http://code.sixapart.com/svn/movabletype/branches/frampton@3097 3860aeb0-bd15-0410-af20-b8d1745f506e