Enabled having multiple config.yaml files per plugin.
This commit adds support for plugin sets with multiple config.yaml files. The convention is this: Melody will consider any file in a plugin directory that ends with "config.yaml" to be a plugin to be loaded. They will then all be associated with the same plugin set.

WARNING and RELEASE NOTE: This is potentially a major change for which there may be backwards compatibility issues. Before this change, plugins loaded via config.yaml had a plugin ID equal to the directory name. This worked because there was only one plugin per directory. However this is no longer the case. Plugins loaded via config.yaml now have a plugin ID that is the conjunction of the directory and their yaml file name. This effectively changes the IDs of plugins which may impact their ability to be loaded properly via MT->component(ID). Additional testing is required.

[#139 state:resolved]