[#362 state:resolved] Tweaked the plugin listing screen so that it automatically selects the first plugin in the list
So what happened here was $app->param("plugin") was not getting set, which screwed up the hashtable.
The next key down, screen_id, was actually being set as the value for plugin in the param hash. Ok, so I lied, that did not quite resolve it. I noticed after I wrote that commit that it was using $app->param in that method instead of $app->query->param so this commit fixes that and brings that method up to spec.

Signed-off-by: Jay Allen <jay@endevver.com>