[#385 state:resolved] Refactored the Six Apart rich text editor into a plugin. This is what I did:
First some background - the "archetype editor" is actually not an editor at all. It is a framework by which one can easily toggle between a plain text editor and a iframe based editor (most likely a rich text editor). The iframe editor is swappable and pluggable by plugins.

So in my refactoring, I pulled out the iframe editor javascript object into its own file, and packaged it into a simple editor plugin. I called it Simple Editor, and attributed the code/plugin to Six Apart. I then made changes to the core to account for when an RTE is NOT installed - because previously it had assumed there was always at least one RTE installed.

This change will allow users for the first time to disable the Six Apart Rich Text Editor from an install all together, and paves the way for us to replace it with one of our (the community's choosing).