[#410 state:resolved] Refactored how plugins and addons are loaded.
In Movable Type the addons directory was assigned a special meaning, and ANY plugin in that directory, regardless of whether it was a "pack" or a "plugin" was loaded before anything in the plugins directory. In Melody this has changed. Here's how:

* Melody begins by scanning the addons directory and all registered directories in the plugins search path.
* A simple list of all packs, plugins and themes (a type not currently utilized) is built. No config.yaml file is parsed. Just a list of config.yaml files to parse later.
* init_addons was changed to consult this list and load ONLY packs. it is called first.
* init_plugins was changed to consult this list and load ONLY plugins. it is called second.

In this change a new bug was found in the Plugin Listing CMS method to keep packs from being surfaced in the Plugin Settings listing screen.