[#515] Fixed bug where users would get "Site URL must be an absolute URL."
This is the result of some outdated logic whereby in order to change the publishing profile of a blog, one would invoke the MT::CMS::Blog::save method/handler. There was a long standing to do to move this to a dedicated handler, which is what we have done here. This bypasses the cms_pre_blog_save_filter or somesuch that was triggering the error. It was safe to bypass because we were not saving the entire blog - we were just updating its profile.