Added a new configuration directive, `AddonPath`, which is an exact analogue to `PluginPath` and is intended to provide feature parity for addons.  See full commit note for further details.
The `PluginPath` config directive allows you to change the name/location of the `plugins` directory or, more importantly, to specify multiple plugin directories.  This is useful if you have multiple installations and want to have a central plugin directory which is used by all. Unfortunately, there has been no way to do the same with the addons directory...until now.