[#889 state:resolved] Fixed bug in which a template's context was clobbered after invoking a template_param callback on a template module include. This resulted in the changes made by a plugin to be lost. It has been an annoying bug for a very, very long time.
The note I made in the bug however is that I am not sure if this original treatment was by design or not. If it was, then my change might introduce a regression I could not anticipate. Recommendation to commiter: let the bug be resolved, but keep a look out for bugs stemming from template_param callbacks, especially on includes. If you get any weird bug reports, assign the case to me for investigation.