[#914] Moved theme managers page_actions to be a child of applications/cms. For some reason this was resulting in the upgrade system to fail because somewhere along the way for mysterious reasons. Well, perhaps not so mysterious, let me explain.
When the Core initialized itself for the purpose of checking the schema, a plugin (in this case Theme Manager) apparently was loading (via a require statement) its CMS.pm file. This module was referencing a schema element that had not yet been initialized and so its compilation was failing, and the whole plugin was not loading, which trickled down to the upgrade framework by reporting that $ThemeManager::ThemeManager::Plugin... was not a HASH element.

By moving page_actions out from the global scope of the Core registry, you prevent the system from trying to load them. Or something.